return message to the user

+++ src/controllers/task.controller.ts
@@ -7,15 +7,16 @@ const taskService = new TaskService();
 
 export class TaskController {
 
-
-
-
     async deleteTsk(req: AuthRequest, res: Response) {
         const task_id = req.params.id;
         if (!task_id || typeof task_id !== 'string') {
             return res.status(400).json({ message: "no task choosen" });
         }
-        const userId = req.currentUser!.id
-        const tasks = taskService.deleteTask(task_id, userId);
+         if (!req.currentUser) {            
+            return res.status(401).json({ message: "Unauthorized" });
+        }
+        const userId = req.currentUser.id;
+        const task =await taskService.deleteTask(task_id, userId);
+        res.status(200).json({message:`task ${task.title} deleted`})
     }
 }
